A complete Java application (OOP)
About vehicles: A regular vehicle is a vehicle that does not have a subscription therefore it is subject to hourly
parking fee. A subscribed vehicle always has a subscription, but it may be invalid (Hint: Notice isValid method
of a subscription). An official vehicle is a vehicle that belongs to public service, i.e. police cars, ambulances,
etc. They do not pay anything for parking; consider them as having permanently valid subscription.
About subscriptions: A subscription is invalid if today’s date is not within the range of the subscription. The
vehicle that is to be associated with a subscription is created in the constructor of the subscription.